xilinx_transceiver: allow higher QPLL frequencies
Currently the QPLL output frequency is limited to the maximum
transceiver line rate assuming the QPLL divider is set to 1. This limits
the achievable output frequencies because some valid combinations of
QPLL frequency and divider are not allowed.

This change removes this restriction by replacing the voltage/speed
grade limit on the QPLL frequency with a check on the lane rate before
the PLLs are configured. Then the QPLLs can use their full range
specified by F_QPLL[01]RANGE in the datasheets. An invalid QPLL
configuration with the divider set to 1 is not possible because of the
previous check on the lane rate.